mv getrows functions to separeted files
